app/requirements.txt
astroid==2.1.0
boto==2.49.0
boto3==1.9.86
botocore==1.12.86
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
gensim==3.7.0
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.16.0
Pillow==5.4.1
python-dateutil==2.7.5
requests==2.21.0
s3transfer==0.1.13
scipy==1.2.0
six==1.12.0
smart-open==1.8.0
torch
torchvision==0.2.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
